refactor(ir): remove `find_first_base_table()` analysis function
